trader/requirements.txt
ccxt
pandas
pytz
fn
cytoolz
pytest
pycodestyle
TA-Lib
aiohttp
